requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi>=2023.7.22
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cssselect==1.3.0
Deprecated==1.2.18
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-SQLAlchemy==3.1.1
fsspec==2025.3.2
gensim==4.3.1
greenlet==3.2.0
huggingface-hub==0.30.2
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.4.2
limits==5.0.0
lxml[html_clean]==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.5
ordered-set==4.1.0
packaging==24.2
pdfminer.six==20221105
pillow==11.2.1
pluggy==1.5.0
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
safetensors==0.5.3
scipy==1.15.2
sgmllib3k==1.0.0
six==1.17.0
smart-open==7.1.0
soupsieve==2.6
SQLAlchemy==2.0.40
sympy==1.13.3
textblob==0.17.1
tinysegmenter==0.3
tldextract==5.2.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.67.1
transformers==4.30.2
typing_extensions==4.13.2
urllib3==2.4.0
uv==0.6.14
Werkzeug==3.1.3
wrapt==1.17.2